add unix2dos split files
