+++ Cargo.toml
@@ -3,6 +3,18 @@ name = "dos2unix"
 version = "0.1.0"
 edition = "2021"
 
+[[bin]]
+name = "dos2unix"
+path = "src/bin/dos2unix.rs"
+
+[[bin]]
+name = "unix2dos"
+path = "src/bin/unix2dos.rs"
+
+[lib]
+name = "dos2unix_lib"
+path = "src/lib.rs"
+
 [dependencies]
 
 [target.'cfg(unix)'.dependencies]
@@ -12,8 +24,8 @@ libc = "0.2.158"
 winapi = { version = "0.3.9", features = ["consoleapi", "handleapi", "processenv", "winbase", "winnt"] }
 
 [profile.release]
-strip = true  # Automatically strip symbols from the binary.
-opt-level = 'z'  # Optimize for size
-lto = true       # Enable Link Time Optimization
-codegen-units = 1 # Reduce parallel code generation units to increase optimization
-panic = 'abort'  # Remove stack unwinding code
+strip = true
+opt-level = 'z'
+lto = true
+codegen-units = 1
+panic = 'abort'
